makes adding and editing part of the backgrounds and bestiary lists
