+++ README.md
@@ -1,40 +1,36 @@
-### Troika PDF Creator
+# Troika Toolkit
 
-Will be a toolkit to generate content for troika! tabletop rpg. For now there are 2 features, one that creates a pdf with a background for the game depending on the data you introduce, and other that walks you through the process of creating a character for the Game, generating a pdf with your character when you're done.
+A web app for the [Troika!](https://troika.uk.com/) tabletop RPG. Lets you create and print character sheets, browse backgrounds and beasts, run a combat initiative tracker, and manage content as an admin.
 
-**If you are familiarized with troika! and you doubt if the creation of characters is following the rules let me tell you explicitly here that the creation of the characters on the present project follows the exact same logic you find in the rule book.**
+Deployed at **[jfdlv.github.io/troika-pdf-creator](https://jfdlv.github.io/troika-pdf-creator)**.
 
-Tentative future features:
+## Features
 
-* Make a visual initiative tracker.
-* Create a backend api so that we don't need to use mocked data and so that the user's can have credentials and  the characaters they create and generate the pdf of old characters if they'd like.
+- **Character Generator** — rolls dice-based stats and randomly assigns a background, then lets you edit possessions and skills before exporting a PDF character sheet.
+- **Backgrounds** — searchable list of all backgrounds with full stats and a Print PDF option. Admins can add and edit entries via modal form.
+- **Bestiary** — searchable list of all beasts with stats, mien table, and special rules. Admins can add and edit entries via modal form.
+- **Initiative Tracker** — setup phase for adding characters and enemies, then a token-draw phase with round management and the Delay rule (6.4).
+- **Saved Characters** — logged-in users can save characters to their account and view or resume them later.
 
+## Tech Stack
 
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+- React 19 + Vite 5
+- Redux Toolkit (global state)
+- React Router v6 with HashRouter (required for GitHub Pages)
+- MUI v5 (dark theme)
+- Firebase 9 — Auth + Firestore
+- jsPDF + jspdf-autotable — PDF generation
+- @hello-pangea/dnd — drag-and-drop inventory sorting
+- SCSS modules per component
 
-## Available Scripts
+## Commands
 
-In the project directory, you can run:
+```bash
+npm start        # dev server → localhost:5173
+npm run build    # production build → dist/
+npm run preview  # preview the production build locally
+```
 
-### `npm start`
+## Deployment
 
-Runs the app in the development mode.<br />
-Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
-
-The page will reload if you make edits.<br />
-You will also see any lint errors in the console.
-
-### `npm test`
-
-Launches the test runner in the interactive watch mode.<br />
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
-
-### `npm run build`
-
-Builds the app for production to the `build` folder.<br />
-It correctly bundles React in production mode and optimizes the build for the best performance.
-
-The build is minified and the filenames include the hashes.<br />
-Your app is ready to be deployed!
-
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
+Hosted on GitHub Pages. CI/CD via `.github/workflows/deploy.yml` — triggers on push to `Refactor_vite_redux_jspdf`. Firebase config is injected via GitHub Actions secrets (`VITE_FIREBASE_*`).
